+++ src/main/java/com/leonardo/apirelatoriovendas/dtos/SaleResponseDTO.java
@@ -1,6 +1,7 @@
 package com.leonardo.apirelatoriovendas.dtos;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -18,7 +19,7 @@ public class SaleResponseDTO implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private Long id;
-    private Double totalValue;
+    private BigDecimal totalValue;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
     private LocalDateTime dateOfSale;
     private Long sellerId;
